refactor(registrar): extract pure helpers to registrar/registrarHelpers.js (Decomp 1)
Strategic Direction 1 — Decomposition (audit §8).

Extracted from RegistrarPanel.jsx lines 27-363 (337 lines) into a new
dedicated helpers module. These functions are pure (no React, no state)
and were safe to extract first because they have no dependencies on
component state or hooks.

Extracted:
- Constants: API_BASE, REGISTRAR_TAB_LABEL_KEYS, REGISTRAR_STATUS_LABEL_KEYS,
  REGISTRAR_RECORD_KINDS
- Style objects: registrarWorkflowHeaderStyle, TitleStyle, MetaStyle,
  ActionsStyle (4 objects, shared between views)
- Contract normalization: normalizeRegistrarContractValue,
  getRegistrarRecordKind, getRegistrarRecordId, getRegistrarRecordRefs,
  getRegistrarSelectionKey, findRegistrarRecordBySelectionKey
- Backend action helpers: hasBackendAction, getRegistrarActionForStatus
- Patient display: hasBackendPatientDisplayContract, normalizePatientGender,
  hasBackendPatientGenderContract
- Wizard queue: resolveWizardQueueEntryId, normalizeWizardQueueAssignment,
  flattenWizardQueueNumbers, buildPostWizardPaymentRow
- Aggregate row: isMultiRecordAggregateRow
- Format: formatPreviewList

Also:
- Removed unused getApiOrigin import from RegistrarPanel.jsx
  (now imported only in registrarHelpers.js)
- Updated RegistrarPanel.contract.test.jsx to read both files via
  readRegistrarSourceTree() helper (contract tests verify presence
  of functions in the source tree, not the orchestrator file)

Stats:
- RegistrarPanel.jsx: 3892 → 3583 lines (−309 lines, −8%)
- registrarHelpers.js: 322 lines (new)
- Tests: 467/467 passing (no regression)
- Vite build: success
